#10 write_log切り出し
